# Optimization: store duplicate rows and columns